+++ apps/api-harmonization/src/blocks/notification-list/notification-list.controller.ts
@@ -1,6 +1,8 @@
 import { Controller, Get, Headers, Query, UseInterceptors } from '@nestjs/common';
 import { LoggerService } from '@o2s/utils.logger';
 
+import { Auth } from '@o2s/framework/modules';
+
 import { AppHeaders } from '@o2s/api-harmonization/utils/headers';
 
 import { URL } from './';
@@ -13,6 +15,7 @@ export class NotificationListController {
     constructor(protected readonly service: NotificationListService) {}
 
     @Get()
+    @Auth.Decorators.Roles({ roles: [Auth.Constants.Roles.USER, Auth.Constants.Roles.ADMIN] })
     getNotificationListBlock(@Headers() headers: AppHeaders, @Query() query: GetNotificationListBlockQuery) {
         return this.service.getNotificationListBlock(query, headers);
     }