+++ apps/api-harmonization/src/utils/date.ts
@@ -5,18 +5,38 @@ import 'dayjs/locale/pl';
 import isToday from 'dayjs/plugin/isToday';
 import isYesterday from 'dayjs/plugin/isYesterday';
 import localizedFormat from 'dayjs/plugin/localizedFormat';
+import timezone from 'dayjs/plugin/timezone';
+import utc from 'dayjs/plugin/utc';
 
 dayjs.extend(isToday);
 dayjs.extend(isYesterday);
 dayjs.extend(localizedFormat);
+dayjs.extend(utc);
+dayjs.extend(timezone);
 
 export const DATE_FORMAT = 'L';
 
-export const formatDateRelative = (date: string | number, locale: string, today: string, yesterday: string) => {
-    const formattedDate = dayjs(date);
+export const formatDateRelative = (
+    date: string | number,
+    locale: string,
+    today: string,
+    yesterday: string,
+    timezone?: string,
+) => {
+    let formattedDate = dayjs(date);
+
+    if (!formattedDate.isValid()) {
+        return '';
+    }
 
     if (locale) {
-        formattedDate.locale(locale);
+        formattedDate = formattedDate.locale(locale);
+    }
+
+    if (timezone) {
+        formattedDate = formattedDate.tz(timezone);
+    } else {
+        formattedDate = formattedDate.tz('Europe/London');
     }
 
     if (formattedDate.isToday()) {