+++ apps/api-harmonization/src/utils/headers.ts
@@ -1,4 +1,6 @@
 export class AppHeaders {
     'x-locale'!: string;
+    'x-client-timezone'?: string;
     'x-currency'?: string;
+    'authorization'?: string;
 }