+++ apps/frontend/next.config.ts
@@ -3,14 +3,9 @@ import type { NextConfig } from 'next';
 // @ts-expect-error missing types for this library
 import withPlugins from 'next-compose-plugins';
 import createNextIntlPlugin from 'next-intl/plugin';
-import withPwa from 'next-pwa';
 
 const withBundleAnalyzer = createBundleAnalyzer({
     enabled: process.env.ANALYZE === 'true',
-});
-
-const withPWA = withPwa({
-    dest: 'public',
 });
 
 const withNextIntl = createNextIntlPlugin();
@@ -32,14 +27,6 @@ const nextConfig: NextConfig = {
         silenceDeprecations: ['legacy-js-api'],
     },
     experimental: {
-        turbo: {
-            rules: {
-                '*.svg': {
-                    loaders: ['@svgr/webpack'],
-                    as: '*.js',
-                },
-            },
-        },
         // dynamicIO: true,
         // cacheLife: {
         //     render: {
@@ -48,6 +35,14 @@ const nextConfig: NextConfig = {
         //         expire: 5,
         //     },
         // },
+    },
+    turbopack: {
+        rules: {
+            '*.svg': {
+                loaders: ['@svgr/webpack'],
+                as: '*.js',
+            },
+        },
     },
     webpack(config) {
         config.module.rules.push({
@@ -59,4 +54,4 @@ const nextConfig: NextConfig = {
     },
 };
 
-export default withPlugins([withPWA, withBundleAnalyzer, withNextIntl], nextConfig);
+export default withPlugins([withBundleAnalyzer, withNextIntl], nextConfig);