+++ apps/frontend/src/blocks/Faq/Faq.server.tsx
@@ -8,13 +8,17 @@ import { FaqProps } from './Faq.types';
 export const FaqDynamic = dynamic(() => import('./Faq.client').then((module) => module.FaqPure));
 
 export const Faq: React.FC<FaqProps> = async ({ id, accessToken, locale }) => {
-    const data = await sdk.blocks.getFaq(
-        {
-            id,
-        },
-        { 'x-locale': locale },
-        accessToken,
-    );
+    try {
+        const data = await sdk.blocks.getFaq(
+            {
+                id,
+            },
+            { 'x-locale': locale },
+            accessToken,
+        );
 
-    return <FaqDynamic {...data} id={id} accessToken={accessToken} locale={locale} />;
+        return <FaqDynamic {...data} id={id} accessToken={accessToken} locale={locale} />;
+    } catch (_error) {
+        return null;
+    }
 };