+++ apps/frontend/src/containers/Header/ContextSwitcher/ContextSwitcher.types.ts
@@ -1,5 +1,0 @@
-import { CMS } from '@o2s/framework/modules';
-
-export interface ContextSwitcherProps {
-    context?: CMS.Model.Header.ContextSwitcher;
-}