+++ apps/frontend/src/containers/Header/DesktopNavigation/DesktopNavigation.tsx
@@ -1,11 +1,10 @@
 'use client';
 
 import { useSession } from 'next-auth/react';
-import { useLocale } from 'next-intl';
+import React from 'react';
 
 import { Models } from '@o2s/framework/modules';
 
-import { Link } from '@o2s/ui/components/link';
 import {
     NavigationMenu,
     NavigationMenuContent,
@@ -32,10 +31,9 @@ export function DesktopNavigation({
     items,
 }: DesktopNavigationProps) {
     const session = useSession();
-    const isSignedIn = session?.status === 'authenticated';
+    const isSignedIn = !!session.data?.user;
 
     const pathname = usePathname();
-    const locale = useLocale();
 
     const activeNavigationGroup = isSignedIn
         ? items.find((item) => {
@@ -73,12 +71,8 @@ export function DesktopNavigation({
         active?: boolean;
     }) => {
         return (
-            <NavigationMenuLink asChild active={active}>
-                <Link asChild>
-                    <NextLink href={href} locale={locale} className={cn(navigationItemClass, className)}>
-                        {children}
-                    </NextLink>
-                </Link>
+            <NavigationMenuLink asChild active={active} className={cn(navigationItemClass, className)}>
+                <NextLink href={href}>{children}</NextLink>
             </NavigationMenuLink>
         );
     };
@@ -152,14 +146,7 @@ export function DesktopNavigation({
                                 {items.map((item) => {
                                     switch (item.__typename) {
                                         case 'NavigationItem':
-                                            return (
-                                                <NavigationItem
-                                                    item={item}
-                                                    key={item.label}
-                                                    active={false}
-                                                    className={'!text-navbar-sub-muted'}
-                                                />
-                                            );
+                                            return <NavigationItem item={item} key={item.label} active={false} />;
                                         case 'NavigationGroup':
                                             return (
                                                 <NavigationItem
@@ -170,7 +157,6 @@ export function DesktopNavigation({
                                                     }}
                                                     key={item.title}
                                                     active={false}
-                                                    className={'!text-navbar-sub-muted'}
                                                 />
                                             );
                                     }
@@ -212,7 +198,7 @@ export function DesktopNavigation({
                                                     item={item}
                                                     key={item.label}
                                                     active={pathname === item.url}
-                                                    className="!text-base"
+                                                    className="!text-base !text-navbar-sub-foreground hover:!text-navbar-sub-foreground hover:!bg-navbar-sub-accent"
                                                 />
                                             );
                                         case 'NavigationGroup':