+++ packages/configs/eslint-config/api.js
@@ -1,25 +1,17 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-    plugins: ['@typescript-eslint/eslint-plugin'],
-    extends: ['plugin:@typescript-eslint/recommended', 'plugin:prettier/recommended'],
-    root: true,
-    env: {
-        node: true,
-        jest: true,
-    },
-    ignorePatterns: ['dist/', '.eslintrc.js', 'lint-staged.config.js'],
-    rules: {
-        '@typescript-eslint/no-unused-vars': [
-            'error',
-            {
-                args: 'all',
-                argsIgnorePattern: '^_',
-                caughtErrors: 'all',
-                caughtErrorsIgnorePattern: '^_',
-                destructuredArrayIgnorePattern: '^_',
-                varsIgnorePattern: '^_',
-                ignoreRestSiblings: true,
-            },
-        ],
-    },
-};
+import js from '@eslint/js';
+import eslintConfigPrettier from 'eslint-config-prettier';
+import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended';
+import tseslint from 'typescript-eslint';
+
+import { config as baseConfig } from './base.js';
+
+/**
+ * @type {import("eslint").Linter.Config[]}
+ * */
+export const config = [
+    ...baseConfig,
+    js.configs.recommended,
+    eslintConfigPrettier,
+    ...tseslint.configs.recommended,
+    eslintPluginPrettierRecommended,
+];