+++ packages/integrations/algolia/.prettierrc.mjs
@@ -1,4 +1,4 @@
-import apiConfig from "@o2s/prettier-config/api.mjs";
+import apiConfig from '@o2s/prettier-config/api.mjs';
 
 /**
  * @see https://prettier.io/docs/en/configuration.html