+++ packages/integrations/mocked/src/modules/cms/cms.service.ts
@@ -15,6 +15,7 @@ import { mapLoginPage } from './mappers/cms.login-page.mapper';
 import { mapNotFoundPage } from './mappers/cms.not-found-page.mapper';
 import { mapNotificationDetailsBlock } from './mappers/cms.notification-details.mapper';
 import { mapNotificationListBlock } from './mappers/cms.notification-list.mapper';
+import { mapOrganizationList } from './mappers/cms.organization-list.mapper';
 import { getAllPages, getAlternativePages, mapPage } from './mappers/cms.page.mapper';
 import { mapPaymentsHistoryBlock } from './mappers/cms.payments-history.mapper';
 import { mapPaymentsSummaryBlock } from './mappers/cms.payments-summary.mapper';
@@ -136,5 +137,9 @@ export class CmsService implements CMS.Service {
 
     getTicketRecentBlock(options: CMS.Request.GetCmsEntryParams) {
         return of(mapTicketRecentBlock(options.locale)).pipe(responseDelay());
+    }
+
+    getOrganizationList(options: CMS.Request.GetCmsOrganizationListParams) {
+        return of(mapOrganizationList(options.locale)).pipe(responseDelay());
     }
 }