+++ packages/integrations/algolia/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @o2s/integrations.algolia
 
+## 0.3.0
+
+### Minor Changes
+
+- de00274: updated dependencies
+
+### Patch Changes
+
+- Updated dependencies [c0ff0a7]
+- Updated dependencies [de00274]
+- Updated dependencies [c0ff0a7]
+- Updated dependencies [e9dc277]
+- Updated dependencies [e9dc277]
+    - @o2s/framework@0.21.0
+    - @o2s/utils.logger@0.10.0
+
 ## 0.2.1
 
 ### Patch Changes